Finalized the way I want users to be added. Added photo upload capabilities for Users. (References #2.)

+++ app/Controller/UsersController.php
@@ -21,44 +21,46 @@ class UsersController extends AppController {
 	 * @return new file url, or a blank string if something went wrong
 	 */
 	private function _uploadFile($data){
-		$photo = $data['User']['picture'];
-		if($photo['name'] == ""){ // Don't reset.
-			return $data['User']['photo'];
-		}
+		if(array_key_exists('picture', $data['User'])){
+			$photo = $data['User']['picture'];
+			if($photo['name'] == ""){ // Don't reset.
+				return $data['User']['photo'];
+			}
 		
-		// Validate the file's type
-		if(!in_array($photo['type'], $this->uploadsFileTypes)){
-			$this->Session->setFlash('You may only upload image files. The rest of your data was saved.');
-			return "";
-		}
+			// Validate the file's type
+			if(!in_array($photo['type'], $this->uploadsFileTypes)){
+				$this->Session->setFlash('You may only upload image files. The rest of your data was saved.');
+				return "";
+			}
 		
-		// Validate the file's size
-		if($photo['size'] > $this->maxFileSize){
-			$this->Session->setFlash('Your upload cannot exceed '. ($this->maxFileSize/1000) .'KB. Please try a smaller photo.');
-			return "";
-		}
+			// Validate the file's size
+			if($photo['size'] > $this->maxFileSize){
+				$this->Session->setFlash('Your upload cannot exceed '. ($this->maxFileSize/1000) .'KB. Please try a smaller photo.');
+				return "";
+			}
 		
-		// Hash filename
-		$filename = substr($photo['name'], 0, strrpos($photo['name'], '.'));
-		$extension = substr($photo['name'], strrpos($photo['name'], '.'));
-		$hashedName = md5($filename) . $extension;
+			// Hash filename
+			$filename = substr($photo['name'], 0, strrpos($photo['name'], '.'));
+			$extension = substr($photo['name'], strrpos($photo['name'], '.'));
+			$hashedName = md5($filename) . $extension;
 		
-		// Create paths
-		$path = DS . $this->uploadsFolder . DS . $hashedName;
-		$fullpath = realpath(dirname(__FILE__) . DS . '..' . DS . 'webroot') . $path;
+			// Create paths
+			$path = DS . $this->uploadsFolder . DS . $hashedName;
+			$fullpath = realpath(dirname(__FILE__) . DS . '..' . DS . 'webroot') . $path;
 		
-		// Insure that directories are there.
-		if(!file_exists(dirname($fullpath))){
-			mkdir(dirname($fullpath), 0777, true);
-		}
+			// Insure that directories are there.
+			if(!file_exists(dirname($fullpath))){
+				mkdir(dirname($fullpath), 0777, true);
+			}
 			
-		// Move files & return file path
-		if(move_uploaded_file($photo['tmp_name'], $fullpath)){
-			$this->Session->setFlash('Your file was uploaded successfully.');
-			return $path;
-		}else{
-			$this->Session->setFlash('Your file could not be uploaded.');
-			return "";
+			// Move files & return file path
+			if(move_uploaded_file($photo['tmp_name'], $fullpath)){
+				$this->Session->setFlash('Your file was uploaded successfully.');
+				return $path;
+			}else{
+				$this->Session->setFlash('Your file could not be uploaded.');
+				return "";
+			}
 		}
 		return "";
 	}
@@ -120,6 +122,10 @@ class UsersController extends AppController {
 	public function add() {
 		if ($this->request->is('post')) {
 			$this->User->create();
+			
+			$this->request->data['User']['photo'] = $this->_uploadFile($this->request->data);
+			$this->request->data['User']['date_created'] = date("Y-m-d H:i:s");
+			
 			if ($this->User->save($this->request->data)) {
 				$this->Session->setFlash(__('The user has been saved'));
 				$this->redirect(array('action' => 'index'));
@@ -143,6 +149,10 @@ class UsersController extends AppController {
 			throw new NotFoundException(__('Invalid user'));
 		}
 		if ($this->request->is('post') || $this->request->is('put')) {
+			
+			$fieldList = array('name', 'show_contact_info', 'email', 'phone_number', 'graduation_year', 'company', 'position', 'bio', 'photo');
+			$this->request->data['User']['photo'] = $this->_uploadFile($this->request->data);
+			
 			if ($this->User->save($this->request->data)) {
 				$this->Session->setFlash(__('The user has been saved'));
 				$this->redirect(array('action' => 'index'));
